dhcpcd: expand set of interfaces initialized
Recognize prefixes to which 'ethX' devices can be renamed.  The
renaming is done by get_ifname_prefix() in udev-builtin-net_id.c
which is in systemd source code.  The list here is from systemd-255.1.
We require at least one character after the prefix, but we don't
look for specific alternatives.

When started at boot time, dhcpcd is run using a wrapper that
feeds the regexes to 'grep -w', which does not recognize constructs
like '(en|wl)'. Thus, we have to list out the alternatives.

Issue: LINCD-13573

Signed-off-by: Joe Slater <joe.slater@windriver.com>