Change RECIPE_LIST_LAYERS to empty string
Before the change, RECIPE_LIST_LAYERS gets the initial value from
BBFILE_COLLECTIONS. It is then overriden by wrlinux-recipelists.inc, and
has a list of Wind River layers. In this case, all receipes in customer
layers are supported without the use of CUSTOMER_RECIPES.

In case RECIPE_LIST_LAYERS is not overriden (e.g. remove require
conf/wrlinux-recipelists.inc), RECIPE_LIST_LAYERS is BBFILE_COLLECTIONS.
It has customer layers. As a result, recipes in customer layers are not
supported if CUSTOMER_RECIPES is not used.

With this change, RECIPE_LIST_LAYERS does not contain any layer names. If
it is not overriden, the checking of whether or not recipes are
supported is skipped, so all recipes in customer layers are supported.

Issue: LINCD-15434

Signed-off-by: Eric Yang <zhangle.yang@windriver.com>