Support mulitilib headers for natveisdk-glibc
This routine will rename the header and add a unique identifier to the name for
the ABI/bitsize that is being used. A wrapper will be generated for the
architecture that knows how to call all of the ABI variants for that given
architecture.

Yocto disable it for class natveisdk in oe-core, this patch support it for
natveisdk-glibc in Wind River Linux

Issue: LINUXEXEC-33826

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>