packagegroup-core-boot-wrs: rdepend on initscripts only in case of sysvinit
Issue: LIN1025-1688

When 'feature/readonly-root' template is enabled, 'bitbake wrlinux-image-initramfs'
fails. This is because this initramfs image is using wrlinux specific packagegroup,
packagegroup-core-boot-wrs.

This packagegroup should does the similar as the oe-core's packagegroup-core-boot,
depending on VIRTUAL-RUNTIME_initscripts only when sysvinit is there.

This readonly regression was caused by the following commit in oe-core:

  "e88f22da19 systemd-compat-units: execute postinst on target"

We need to sync with oe-core's packagegroup-core-boot's behavior to avoid error.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>