apply-projectinfo.py: improve it
Issue: LIN1025-6201

Adding four new options:

--base-url / --base-branch:
These options replace the URL and branch in the tarball. They are useful
when the original URL uses HTTPS and the user applying the tarball does
not know the required username or password. In this case, the user can
substitute a different URL instead.

--user / --password:
These options allow the user to pass a username and password to
setup.sh. However, the user will still need to manually enter the
credentials during the wrlinux-x clone process as what they done during
normal setup.

Signed-off-by: Changqing Li <changqing.li@windriver.com>